Migrated old group menu items to new group button
